app/backend/requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.0
Pillow>=10.0.0